backend/requirements.txt
fastapi
uvicorn
beanie
motor
python-dotenv
httpx
anthropic
openai
python-multipart
email-validator
pydantic